update module API
