+++ test/spec.js
@@ -1,5 +1,5 @@
 var assert = require('assert');
-var load = require('../');
+var datadir = require('../');
 var path = require('path');
 
 var createProxy = require('../proxy');
@@ -9,7 +9,7 @@ var localDataPath = path.join(__dirname, '_data');
 describe('load()', function() {
 
   it('reads data', function(done) {
-    load(localDataPath, function(error, data) {
+    datadir.load(localDataPath, function(error, data) {
       assert.deepEqual(data.test, {foo: 'bar'});
       done();
     });
@@ -20,7 +20,7 @@ describe('load()', function() {
 describe('proxy', function() {
 
   it('creates proxies that read data', function() {
-    var data = createProxy(localDataPath);
+    var data = datadir.proxy(localDataPath);
     assert.deepEqual(data.test, {foo: 'bar'});
   });
